+++ common/src/main/java/mod/adrenix/nostalgic/mixin/tweak/gameplay/combat_item/DiggerItemMixin.java
@@ -71,9 +71,11 @@ public abstract class DiggerItemMixin extends TieredItem
     )
     private Multimap<Attribute, AttributeModifier> nt_combat_item$modifyDiggerItemDefaultAttributeModifiers(Multimap<Attribute, AttributeModifier> defaultModifiers, EquipmentSlot slot)
     {
-        if (GameplayTweak.OLD_DAMAGE_VALUES.get() && EquipmentSlot.MAINHAND == slot && this.nt$isModifierAllowed)
+        // @formatter:off
+        if (GameplayTweak.OLD_DAMAGE_VALUES.get() && EquipmentSlot.MAINHAND == slot && this.nt$isModifierAllowed && !GameplayTweak.IGNORED_OLD_DAMAGE_ITEMS.get().containsItem(this))
             return this.nt$defaultModifiers;
 
+        // @formatter:on
         return defaultModifiers;
     }
 }