+++ common/src/main/java/mod/adrenix/nostalgic/mixin/tweak/sound/game_music/MusicManagerMixin.java
@@ -2,11 +2,15 @@ package mod.adrenix.nostalgic.mixin.tweak.sound.game_music;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import mod.adrenix.nostalgic.helper.sound.MusicHelper;
+import mod.adrenix.nostalgic.tweak.config.SoundTweak;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.sounds.SoundInstance;
 import net.minecraft.client.sounds.MusicManager;
 import net.minecraft.resources.ResourceLocation;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
@@ -15,6 +19,26 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(MusicManager.class)
 public abstract class MusicManagerMixin
 {
+    @Shadow private int nextSongDelay;
+    @Shadow @Final private Minecraft minecraft;
+    @Shadow @Nullable private SoundInstance currentMusic;
+
+    /**
+     * Removes the delay until the next song if music is to play continuously.
+     */
+    @Inject(
+        method = "tick",
+        at = @At("HEAD")
+    )
+    private void nt_game_music$skipMusicDelay(CallbackInfo callback)
+    {
+        if (this.currentMusic == null && this.minecraft.level != null && SoundTweak.PLAY_MUSIC_CONTINUOUSLY.get())
+        {
+            if (this.nextSongDelay > SoundTweak.CONTINUOUS_DELAY.get() * 20)
+                this.nextSongDelay = SoundTweak.CONTINUOUS_DELAY.get() * 20;
+        }
+    }
+
     /**
      * Instructs the music manager to also check if our music is still playing.
      */
@@ -30,7 +54,7 @@ public abstract class MusicManagerMixin
         if (MusicHelper.CURRENT_SONG.isEmpty())
             return isMusicActive;
 
-        if (Minecraft.getInstance().getSoundManager().isActive(MusicHelper.CURRENT_SONG.getOrThrow()))
+        if (this.minecraft.getSoundManager().isActive(MusicHelper.CURRENT_SONG.getOrThrow()))
             return true;
         else
             MusicHelper.CURRENT_SONG.clear();
@@ -70,7 +94,7 @@ public abstract class MusicManagerMixin
     )
     private void nt_game_music$stopC418Music(CallbackInfo callback)
     {
-        MusicHelper.CURRENT_SONG.ifPresent(song -> Minecraft.getInstance().getSoundManager().stop(song));
+        MusicHelper.CURRENT_SONG.ifPresent(song -> this.minecraft.getSoundManager().stop(song));
         MusicHelper.CURRENT_SONG.clear();
     }
 }